third_party/blink/web_tests/platform/generic/fast/media/mq-pixel-ratio-print-expected.txt

CSS3 media query test: -webkit-device-pixel-ratio combined with print.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matchMedia('(-webkit-min-device-pixel-ratio: 1)').matches is true
PASS resolutionFromStyle() is 1
PASS matchMedia('(-webkit-min-device-pixel-ratio: 3)').matches is true
PASS resolutionFromStyle() is 3
PASS resolutionFromStyle() is 3.125
PASS successfullyParsed is true

TEST COMPLETE
